+++ Sources/SafePreviewDevice/ViewExtension.swift
@@ -10,8 +10,20 @@ import SwiftUI
 
 @available(iOS 13.0, OSX 10.15, tvOS 13.0, watchOS 6.0, *)
 public extension View {
+    
+    /// Overrides the device for a preview.
+    /// - Parameter device: the choosen device
+    /// - Returns: The Xcode preview canvas `View`
     func previewDevice(_ device: Devices) -> some View {
-        return PreviewDevice(device: device)
+        let device: PreviewDevice? = PreviewDevice(device: device)
+        return previewDevice(device)
+    }
+    
+    /// Provides a user visible name shown in the editor.
+    /// - Parameter device: the choosen device
+    /// - Returns: The visible name label below preview canvas
+    func previewDeviceName(_ device: Devices) -> some View {
+        previewDisplayName(device.rawValue)
     }
 }