+++ CHANGELOG.md
@@ -4,11 +4,14 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.1.0] Hello Release 👋
 
 ### Added
 
-- None
+- Two `View` extensions:
+    `previewDevice(_ device: Devices)`
+    `previewDeviceName(_ device: Devices)`
+    that will semplify your PreviewDevice choice.
 
 ### Changed